BACKEND/requirements.txt
pandas
openai
python-dotenv
opencv-python
numpy
openai
PyMuPDF
tensorflow
python-doctr
huggingface_hub
fastapi
uvicorn[standard]
python-multipart
tqdm
